kola: Write ignition-virtio-dump.txt to test dir
We are seeing the `rhcos.fips` test fail occasionally; the
console is the usual mess of corrupted/overwritten text.

We have the new virtio dump stuff, but until now we have
been discarding the output.  Change the test framework to
save it.

(And in the future we should parse it and try to find
 actual error messages)