mantle/ore: glcoud: add --create-image option to upload.go
In RHCOS we upload images to GCP but no one ever uses them because
the openshift-installer just downloads the image from the URL in the
metadata and then recreates the image from scratch. Add the
--create-image boolean option (defaults to true) so we can specify
`--create-image=false` for the RHCOS pipeline and stop wasting time
and resources doing the image creation.