cosalib/gcp: attach to image family in separate API call
Currently we create an image and attach it to an image family in a
single API call. We'd like to attach it to an image family *after*
the image has been deprecated (deprecated images won't be given
to users who are trying to retrieve the latest image for an image
family). However, you can't create an image AND deprecate it in
the same API call so there is a split second of opportunity for
a user to get the newer GCP image before we've officially released
it.

This change breaks it up so that we can:

1. create the image (not part of a image family)
2. deprecate the image
3. attach the image to an image family

The new flow means there is no opportunity for a user to get an
image that has yet to be released.