create_disk: Add known UUIDs for /boot and /root
Right now we allocate a random UUID on each build,
but everyone starting from a particular build will have
the same one...so they're not unique.

This will pair with a new version of
https://github.com/coreos/ignition-dracut/pull/183
so that we can regenerate them on firstboot and ensure
that they are actually unique (as much as possible)
across running systems.