upload-oscontainer: Automatically inject --display-name
At one point I was trying to drag the oscontainer code
out of the RHCOS pipeline, but stalled on that.
While I was doing that, we added `--display-name` logic
in the downstream version.

When I went to switch everything to upstream, that
change got lost.

The previous downstream change hardcoded passing
`--display-name=<osname>` in the pipeline.
Let's do better and automatically extract `os-release`
and use it.