upload-oscontainer: Correctly strip quotes from name
The previous change to automatically add the display name
appeared to work when I tested locally, and I thought the
quotes in the output of `podman inspect` were added by podman
itself.

But since the file is in "shell syntax", it can have shell-style
quotes, so let's use the shell to parse it.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1847886