README.md: Add a note about `--net=host` here.
This came from https://github.com/coreos/coreos-assembler/pull/282
which actually we don't strictly need for builds anymore because
we rely on 9p.

But some people may want it for e.g. kola.  It is tempting to
make `--net=host` the default particularly because SLIRP is a
performance hit but let's just document this for now.