mantle: Paper over another reboot failure mode
This one isn't a new bug I introduced, I swear!  The ostree
test suite seems to hit this more often for some reason
I didn't fully debug though.

As I note in the comment I think we can clean this up in a
much better way in the future, but for now I'd like to paper
over this one and move on to getting OSTree's CI going again.