buildextend-installer: move PXE squashfs and osmet to separate initrd
For live images, add PXE "rootfs" image containing osmet files and the root
squashfs.  This can be appended as a second initrd or fetched by the
initramfs at runtime.  Write the build ID into flag files in the initramfs
and rootfs so the initramfs can confirm that both came from the same build.
Store a chunked hash of the rootfs into the PXE initramfs so rootfs
integrity can be verified after a runtime fetch.

If --legacy-pxe is specified, continue as we were before, but generate a
rootfs image containing only a flag file.  This allows a transition period
where the initramfs can check for the flag file and warn that the rootfs
hasn't been provided.