mantle: gcp: add ability to make gcp images public
This adds support for marking images with public ACLs when they are
created. Doing this isn't necessary for projects that have been marked
as public on the GCP backend by GCP engineers (like the `fedora-coreos-cloud`
project), but is for other projects to share images publicly.

Fixes https://github.com/coreos/coreos-assembler/issues/1561