mantle/system/exec: allow parallel Wait() and Kill()
Kill() is just canceling a Context, so it's safe after the process exits.
We just need to make sure its associated Wait() call doesn't cause
exec.Cmd.Wait() to be called more than once.  Add a wrapper so Wait()
can be called any number of times.

This helps callers that want to Wait() from a goroutine and Kill() after
a timeout, which is otherwise difficult to do correctly.  In particular,
QemuInstance was trying to coordinate by assigning nil to inst.qemu in
a non-thread-safe way.