qemu: Automatically inject serial console karg into ISO
Now that we support injecting kernel arguments in the ISO, let's make
use of it by default when booting a live ISO to inject the appropriate
serial console karg. This is more convenient for manual testing (because
one no longer has to interrupt the boot process to type it in), but also
has the major advantage of ISO tests now being able to capture console
output.

In the future, we can also hook up the `--kargs` option to this.