oscontainer: Use skopeo, not podman to extract
I kept running into failures trying to run this in an
unprivileged container; basically the containers stack
seems to want to try to `chown()` files based on the UIDs,
but we don't care about that at all for this.

What we really want is the functionality of `oc image extract`
but that doesn't exist yet in the podman stack and today
cosa doesn't include `oc` (that's a big discussion).