Add an image-default.yaml, pass JSON to create_disk.sh
The way we parse an `image.yaml` and then convert it
into arguments for `create_disk.sh` is unwieldy.

Instead, convert `image.yaml` into JSON, inheriting default values
from a separate `image-default.yaml` we ship.  Then pass
that JSON to `create_disk.sh`.

This only converts the `rootfs` argument, but if people
are OK with this I can go ahead and convert everything
else - getting us to where the only argument to `create_disk.sh`
is a JSON file that's already been pre-validated.