Write out a ref with the build ID
I'm not sure why we didn't do this from the start.  The immediate
motivation here is I was trying to diff two builds and the ostree
commit previously was pruned while the cosa build still existed.

Now yes I could import the ostree tarball but this seems clearly
better.

With this we can also drop the `tmpref-` hack we have for RHCOS
and just use the build ID as a ref.