Use build ref or commit, drop "ref_is_tmp" hack
This is a general cleanup followup to the previous PR
to add a "buildid" as a ref always.
In most cases actually where we were using the ref we can
just use the commit.

But switch a few things to use the buildid, and drop the
`ref_is_tmp` hack.