kola: fix upgrade tests handling cosa build
The comment says that `--build` is required, but what we really mean is
that a cosa build is required. Whether that build is auto-discovered or
given explicitly via `--build` doesn't matter. Move the check for a cosa
build down to after `syncOptionsImpl()` to reflect this.

This also fixes a missing `return` which Jing originally had in #1826.

Co-authored-by: Jing Zhang <jingzhan@redhat.com>
Closes: #1827