create_disk: Also xfs_freeze /boot
Followup to ca97e50c6de870664f576f9427f377749a2ee19e
This doesn't change anything today because `/boot` is ext4
which flushes the journal on unmount.  But I was just reading
this script and noticed that we should be doing the same thing
for `/boot` that we do for root.